backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "datetime>=6.0",
    "dotenv>=0.9.9",
    "elevenlabs>=1.0.0",
    "fastapi>=0.135.1",
    "google-genai>=1.0.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pymongo>=4.16.0",
    "python-multipart>=0.0.9",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
]

